+++ RayTracerChallenge.Domain/Extensions/MatrixExtensions.cs
@@ -4,13 +4,13 @@ public static class MatrixExtensions
 {
     private static TransformationFactory _transformationFactory = new TransformationFactory();
 
-    public static Matrix Translate(this Matrix m, float x, float y, float z)
+    public static Matrix Translate(this Matrix m, decimal x, decimal y, decimal z)
     {
         var translate = _transformationFactory.Translation(x, y, z);
         return translate * m;
     }
 
-    public static Matrix Scale(this Matrix m, float x, float y, float z)
+    public static Matrix Scale(this Matrix m, decimal x, decimal y, decimal z)
     {
         var scale = _transformationFactory.Scale(x, y, z);
         return scale * m;
@@ -34,9 +34,9 @@ public static class MatrixExtensions
         return rotation * m;
     }
 
-    public static Matrix Shear(this Matrix m, float xToY, float xToZ,
-        float yToX, float yToZ,
-        float zToX, float zToY)
+    public static Matrix Shear(this Matrix m, decimal xToY, decimal xToZ,
+        decimal yToX, decimal yToZ,
+        decimal zToX, decimal zToY)
     {
         var shear = _transformationFactory.Shearing(xToY, xToZ, yToX, yToZ,
             zToX, zToY);
